+++ src/main/java/burgerbuilder/dao/CustomerRepository.java
@@ -7,7 +7,6 @@ import org.springframework.data.repository.CrudRepository;
 import burgerbuilder.domain.Customer;
 
 public interface CustomerRepository extends CrudRepository<Customer, Integer> {
-	Iterable<Optional<Customer>> findByCustomerId(Integer customerId);
 	Iterable<Optional<Customer>> findByAddress(Integer address);
 	Iterable<Optional<Customer>> findByEmail(String email);
 	Iterable<Optional<Customer>> findByName(String name);
